moved github event hook to auth
